ai-service/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
langchain-openai
faiss-cpu
python-multipart
openai-whisper